remove optional GSF dependency
